ml benchmark dockerfiles (#1)
* ml benchmark dockerfiles

In preparation of testing and understanding ML apps,
I am preparing the docker containers. I was hoping to test
on V100 on Google Cloud, but I am also wanting to update
to CUDA 12.8 and unsure if it will work. We will find out.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>

* ml benchmark containers

These are a basic set of ML benchmarks
we can test, likely first on Google Cloud
(older GPU) and then move over to AWS.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>

---------

Signed-off-by: vsoch <vsoch@users.noreply.github.com>
Co-authored-by: vsoch <vsoch@users.noreply.github.com>